refactor(suite): move Address into @suite/address
