+++ packages/suite/src/components/earn/dashboard/yield/EarnYieldClaimSelectAccountModal.tsx
@@ -1,3 +1,4 @@
+import { Address } from '@suite/address';
 import { events } from '@suite/analytics';
 import { Translation } from '@suite/intl';
 import { useFormatters } from '@suite-common/formatters';
@@ -5,7 +6,6 @@ import { CardList, Column, Modal, Row, Text } from '@trezor/components';
 import { CoinLogo } from '@trezor/product-components';
 
 import { AccountLabel } from 'src/components/suite/AccountLabel';
-import { Address } from 'src/components/suite/Address';
 import { useAnalytics } from 'src/support/useAnalytics';
 
 import { type YieldAccountRewards, type YieldAccountsRewards } from '../../yield/claim/hooks';
