+++ packages/suite/src/components/suite/index.tsx
@@ -5,7 +5,6 @@ WARNING - do NOT import from this file in the suite/src/components/suite/ subdir
 /* eslint-disable import/order */
 // TODO Change this to direct export {} from, instead of importing and re-exporting, but currently cannot be done because of circular dependencies.
 import { AccountLabel } from './AccountLabel';
-import { Address } from './Address';
 import { DeviceConfirmImage } from './DeviceConfirmImage';
 import { CheckItem } from './CheckItem';
 import { FakeSelect } from './FakeSelect';
@@ -57,7 +56,6 @@ import { MarkdownWithComponents } from './MarkdownWithComponents';
 import { AppRouter } from './AppRouter';
 
 export {
-    Address,
     AccountLabel,
     DeviceConfirmImage,
     CheckItem,
