+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/TxDetailModal/AffectedTransactions/DecreasedOutputs.tsx
@@ -1,3 +1,4 @@
+import { Address } from '@suite/address';
 import { Translation, type TranslationKey } from '@suite/intl';
 import { type NetworkSymbol } from '@suite-common/wallet-config';
 import { type Account, type FormState } from '@suite-common/wallet-types';
@@ -16,7 +17,6 @@ import {
 import { spacings } from '@trezor/theme';
 import { HELP_CENTER_REPLACE_BY_FEE_BITCOIN } from '@trezor/urls';
 
-import { Address } from 'src/components/suite/Address';
 import { FormattedCryptoAmount } from 'src/components/suite/FormattedCryptoAmount';
 import { HiddenPlaceholder } from 'src/components/suite/HiddenPlaceholder';
 import { type RbfContextValues, useRbfContext } from 'src/hooks/wallet/useRbfForm';
