+++ packages/suite/src/components/tx-simulation/common/components/TxSimulationContractInfo.tsx
@@ -1,10 +1,10 @@
+import { Address } from '@suite/address';
 import { Translation } from '@suite/intl';
 import { type TransactionSimulation } from '@suite-common/tx-simulation';
 import { type Network, getExplorerUrl } from '@suite-common/wallet-config';
 import { selectExplorer } from '@suite-common/wallet-core';
 import { CollapsibleBox, Column, H4, Link, Row, Text } from '@trezor/components';
 
-import { Address } from 'src/components/suite/Address';
 import { useExternalLink, useSelector } from 'src/hooks/suite';
 import { getTokenAddressTranslationId } from 'src/utils/wallet/tokenUtils';
