+++ packages/suite/src/components/wallet/TransactionItem/TransactionTarget/TargetAddressLabel.tsx
@@ -1,3 +1,4 @@
+import { Address } from '@suite/address';
 import { Translation } from '@suite/intl';
 import { selectIsLegacyLabelingVisible } from '@suite/metadata';
 import { type AccountLabels } from '@suite-common/metadata-types';
@@ -8,7 +9,7 @@ import type { NetworkSymbol } from '@suite-common/wallet-config';
 import type { StaticSessionId } from '@trezor/connect';
 import { type ArrayElement } from '@trezor/type-utils';
 
-import { Address, AddressLabeling } from 'src/components/suite';
+import { AddressLabeling } from 'src/components/suite';
 import { useSelector } from 'src/hooks/suite';
 import { type WalletAccountTransaction } from 'src/types/wallet';
