+++ suite/address/package.json
@@ -1,0 +1,34 @@
+{
+    "name": "@suite/address",
+    "version": "1.0.0",
+    "private": true,
+    "license": "See LICENSE.md in repo root",
+    "sideEffects": false,
+    "main": "src/index",
+    "scripts": {
+        "depcheck": "yarn g:depcheck",
+        "lint:js": "yarn g:eslint '**/*.{ts,tsx,js}'",
+        "type-check": "yarn g:tsc --build"
+    },
+    "dependencies": {
+        "@suite-common/device": "workspace:*",
+        "@suite-common/formatters": "workspace:*",
+        "@suite-common/suite-constants": "workspace:*",
+        "@suite-common/toast-notifications": "workspace:*",
+        "@suite-common/wallet-core": "workspace:*",
+        "@suite-common/wallet-types": "workspace:*",
+        "@suite/modal": "workspace:*",
+        "@trezor/components": "workspace:*",
+        "@trezor/device-utils": "workspace:*",
+        "@trezor/dom-utils": "workspace:*",
+        "@trezor/theme": "workspace:*",
+        "react": "19.2.4",
+        "react-redux": "9.2.0",
+        "redux": "^5.0.1",
+        "redux-thunk": "^3.1.0",
+        "styled-components": "^6.3.9"
+    },
+    "devDependencies": {
+        "@types/react": "19.2.14"
+    }
+}
