+++ suite/address/src/index.ts
@@ -1,0 +1,2 @@
+export * from './Address';
+export * from './copyAddressActions';
