+++ suite/address/tsconfig.json
@@ -1,0 +1,29 @@
+{
+    "extends": "../../tsconfig.base.json",
+    "compilerOptions": { "outDir": "libDev" },
+    "references": [
+        { "path": "../../suite-common/device" },
+        {
+            "path": "../../suite-common/formatters"
+        },
+        {
+            "path": "../../suite-common/suite-constants"
+        },
+        {
+            "path": "../../suite-common/toast-notifications"
+        },
+        {
+            "path": "../../suite-common/wallet-core"
+        },
+        {
+            "path": "../../suite-common/wallet-types"
+        },
+        { "path": "../modal" },
+        { "path": "../../packages/components" },
+        {
+            "path": "../../packages/device-utils"
+        },
+        { "path": "../../packages/dom-utils" },
+        { "path": "../../packages/theme" }
+    ]
+}
